Bug fixes
Fixed a bug where wei value wasn't comparing correctly causing non-valued transactions to be added to the database.
Added block id as a database field for lesser API calls during the loop. The most recent block id is now added to the web3 filter versus starting at 0 each time so that significantly less total calls are made.
Compressed code so that it is easier to read
Added TX link to Telegram message